Better encapsulate statement splitting
Hide the implementation details that it is implemented by regular
expressions to allow easier implementation of different techniques in
the future. Also a note on the terminology, it is probably better named
batch splitting, but I wanted to keep the implementation as close to how
it was originally; hence the same naming.

Furthermore, dependent code doesn't care *how* the statements are split,
just that they are - so this further encapsulates that decision and is
a better abstraction point.

Notably some tests have changed from using raw string literals, instead
being parameterized to accept the line endings in different environments
(i.e. Windows and everything else). This helps the tests be more
consistent. (Thanks to @wokket for pointing this issue out:
https://github.com/grate-devs/grate/pull/816#discussion_r3844060141
)
