+++ unittests/Oracle/TestInfrastructure/OracleSplitterContext.cs
@@ -99,25 +99,25 @@ INSERT [dbo].[Foo] ([Bar]) VALUES (N'/ speed racer, / speed racer, / speed racer
 
 /";
 
-        public static readonly string PLSqlStatementScrubbed = @"
+        public static readonly string[] PLSqlStatementScrubbed = [@"
 BOB1
-" + StatementSplitter.BatchTerminatorReplacementString + @"
+", @"
 
 /* COMMENT */
 BOB2
-" + StatementSplitter.BatchTerminatorReplacementString + @"
+", @"
 
 -- /
 
-BOB3 " + StatementSplitter.BatchTerminatorReplacementString + @"
+BOB3 ", @"
 
 --`~!@#$%^&*()-_+=,.;:'""[]\/?<> /
 
 BOB5
-   " + StatementSplitter.BatchTerminatorReplacementString + @"
+   ", @"
 
 BOB6
-" + StatementSplitter.BatchTerminatorReplacementString + @"
+", @"
 
 /* / */
 
@@ -132,12 +132,12 @@ BOB7
 BOB8
 
 --
-" + StatementSplitter.BatchTerminatorReplacementString + @"
+", @"
 
 BOB9
 
 -- `~!@#$%^&*()-_+=,.;:'""[]\/?<>
-" + StatementSplitter.BatchTerminatorReplacementString + @"
+", @"
 
 BOB10/
 
@@ -165,7 +165,7 @@ ALTER TABLE Inv.something ADD
 	slsald varchar(15) NULL,
 	uhasdf varchar(15) NULL,
     daf_asdfasdf DECIMAL(20,6) NULL;
-" + StatementSplitter.BatchTerminatorReplacementString + @"
+", @"
 
 EXEC @ReturnCode = msdb.dbo.sp_add_jobstep @job_id=@jobId, @step_name=N'Daily job', 
 		@step_id=1, 
@@ -182,40 +182,14 @@ dml statements
 /  
 dml statements '
 
-" + StatementSplitter.BatchTerminatorReplacementString + @"
+", @"
 
 INSERT [dbo].[Foo] ([Bar]) VALUES (N'hello--world.
 Thanks!')
 INSERT [dbo].[Foo] ([Bar]) VALUES (N'/ speed racer, / speed racer, / speed racer /!!!!! ')
 
-" + StatementSplitter.BatchTerminatorReplacementString + @"";
+"];
 
-        public static readonly string plsql_statement =
-            @"
-SQL1;
-;
-SQL2;
-;
-tmpSql := 'DROP SEQUENCE mutatieStockID';
-EXECUTE IMMEDIATE tmpSql; 
-;
-BEGIN
-INSERT into Table (columnname) values ("";"");
-UPDATE Table set columnname="";"";
-END;
-";
-        public static readonly string plsql_statement_scrubbed = @"
-SQL1;
-" + StatementSplitter.BatchTerminatorReplacementString + @"
-SQL2;
-" + StatementSplitter.BatchTerminatorReplacementString + @"
-tmpSql := 'DROP SEQUENCE mutatieStockID';
-EXECUTE IMMEDIATE tmpSql; 
-" + StatementSplitter.BatchTerminatorReplacementString + @"
-BEGIN
-INSERT into Table (columnname) values ("";"");
-UPDATE Table set columnname="";"";
-END;
-";
+
     }
 }
